+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 4.6 (Unreleased)
 
+**Design**
+
+- Re-stylize the meta block
+
+
 ## 4.5
 
 This is the largest release since `4.0`, and the biggest feature is dual-pane mode. Dual-pane mode splits the front page